LoadContitionColors -> LoadActivityColors
Replacing function names in all R scripts currently in use. This is to disambiguate "condition" from activity manipulations (LoadActivityColors) and circadian timepoints (LoadZTColors).

May be residual calls to LoadConditionColors in other scripts not in use but still in the repo, mainly in `junk_drawer`. Should throw an error if ever called.